[layer] Update layer cpp API
Update layer cpp API to match the c-api interface.
Move pluggableLayer related stuff to layer_internal.

See Also #986

**Self evaluation:**
1. Build test: [x]Passed [ ]Failed [ ]Skipped
2. Run test: [x]Passed [ ]Failed [ ]Skipped
