+++ api/ccapi/include/layer.h
@@ -64,160 +64,56 @@ public:
   virtual ~Layer() = default;
 
   /**
-   * @brief     set Property of layer
-   * @param[in] values values of property
-   * @retval #ML_ERROR_NONE Successful.
-   * @retval #ML_ERROR_INVALID_PARAMETER invalid parameter.
-   */
-  virtual int setProperty(std::vector<std::string> values) = 0;
-
-  /**
-   * @brief     Property Enumeration
-   *            0. input shape : string
-   *            1. bias zero : bool
-   *            2. normalization : bool
-   *            3. standardization : bool
-   *            4. activation : string (type)
-   *            5. epsilon : float
-   *            6. weight_regularizer : string (type)
-   *            7. weight_regularizer_constant : float
-   *            8. unit : int
-   *            9. weight_initializer : string (type)
-   *            10. filter_size : int
-   *            11. kernel_size : ( n , m )
-   *            12. stride : ( n, m )
-   *            13. padding : ( n, m )
-   *            14. pool_size : ( n,m )
-   *            15. pooling : max, average, global_max, global_average
-   *            16. flatten : bool
-   *            17. name : string (type)
-   *            18. num_inputs : unsigned int (minimum 1)
-   *            19. num_outputs : unsigned int (minimum 1)
-   *            20. momentum : float,
-   *            21. moving_mean_initializer : string (type),
-   *            22. moving_variance_initializer : string (type),
-   *            23. gamma_initializer : string (type),
-   *            24. beta_initializer" : string (type)
-   *            25. modelfile : model file for loading config for backbone layer
-   *            26. input_layers : string (type)
-   *            27. output_layers : string (type)
-   *            28. trainable :
-   *            29. flip_direction
-   *            30. random_translate
-   *            31. in_dim : int ( input dimension for embedding layer )
-   *            32. out_dim : int ( output dimesion for embedding layer )
-   *            33. in_length : int ( input length for embedding layer )
-   *            34. recurrent_activation :  string (type) - lstm
-   */
-  enum class PropertyType {
-    input_shape = 0,
-    normalization = 1,
-    standardization = 2,
-    activation = 3,
-    epsilon = 4,
-    weight_regularizer = 5,
-    weight_regularizer_constant = 6,
-    unit = 7,
-    weight_initializer = 8,
-    bias_initializer = 9,
-    filters = 10,
-    kernel_size = 11,
-    stride = 12,
-    padding = 13,
-    pool_size = 14,
-    pooling = 15,
-    flatten = 16,
-    name = 17,
-    num_inputs = 18,
-    num_outputs = 19,
-    momentum = 20,
-    moving_mean_initializer = 21,
-    moving_variance_initializer = 22,
-    gamma_initializer = 23,
-    beta_initializer = 24,
-    modelfile = 25, /** model file for loading config for backbone layer */
-    input_layers = 26,
-    output_layers = 27,
-    trainable = 28,
-    flip_direction = 29,
-    random_translate = 30,
-    in_dim = 31,
-    out_dim = 32,
-    in_length = 33,
-    recurrent_activation = 34,
-    unknown
-  };
-
-  /**
-   * @brief setProperty by PropertyType
-   * @note By passing empty string, this can validate if @a type is valid
-   * @param[in] type property type to be passed
-   * @param[in] value value to be passed, if empty string is passed, do nothing
-   * but throws error when @a type is invalid
-   * @exception exception::not_supported     when property type is not valid for
-   * the particular layer
-   * @exception std::invalid_argument invalid argument
-   * @note A layer need not support all the properties from PropertyType, but
-   * the supported properties will be a subset of PropertyType.
-   */
-  virtual void setProperty(const PropertyType type,
-                           const std::string &value = "") = 0;
-
-  /**
-   * @brief     check hyper parameter for the layer
-   * @retval #ML_ERROR_NONE Successful.
-   * @retval #ML_ERROR_INVALID_PARAMETER invalid parameter.
-   */
-  virtual int checkValidation() = 0;
-
-  /**
-   * @brief  get the loss value added by this layer
-   * @retval loss value
-   */
-  virtual float getLoss() = 0;
-
-  /**
-   * @brief     set trainable for this layer
-   * @param[in] train to enable/disable train
-   */
-  virtual void setTrainable(bool train) = 0;
-
-  /**
-   * @brief     get if the output of this layer must be flatten
-   * @retval    flatten value
-   */
-  virtual bool getFlatten() = 0;
-
-  /**
-   * @brief     Get name of the layer
-   */
-  virtual std::string getName() noexcept = 0;
-
-  /**
    * @brief Get the layer type
    * @return const std::string type representation
    */
   virtual const std::string getType() const = 0;
 
   /**
-   * @brief Preset modes for printing summary for the layer
+   * @brief     Default allowed properties
+   * - input shape : string
+   * - bias zero : bool
+   * - normalization : bool
+   * - standardization : bool
+   * - activation : string (type)
+   * - epsilon : float
+   * - weight_regularizer : string (type)
+   * - weight_regularizer_constant : float
+   * - unit : int
+   * - weight_initializer : string (type)
+   * - filter_size : int
+   * - kernel_size : ( n , m )
+   * - stride : ( n, m )
+   * - padding : ( n, m )
+   * - pool_size : ( n,m )
+   * - pooling : max, average, global_max, global_average
+   * - flatten : bool
+   * - name : string (type)
+   * - num_inputs : unsigned int (minimum 1)
+   * - num_outputs : unsigned int (minimum 1)
+   * - momentum : float,
+   * - moving_mean_initializer : string (type),
+   * - moving_variance_initializer : string (type),
+   * - gamma_initializer : string (type),
+   * - beta_initializer" : string (type)
+   * - modelfile : model file for loading config for backbone layer
+   * - input_layers : string (type)
+   * - output_layers : string (type)
+   * - trainable :
+   * - flip_direction
+   * - random_translate
+   * - in_dim : int ( input dimension for embedding layer )
+   * - out_dim : int ( output dimesion for embedding layer )
+   * - in_length : int ( input length for embedding layer )
+   * - recurrent_activation : string (type) - used only in lstm
    */
-  enum class PrintPreset {
-    PRINT_NONE = 0,     /**< Print nothing */
-    PRINT_SUMMARY,      /**< Print preset including summary information */
-    PRINT_SUMMARY_META, /**< Print summary preset that includes meta information
-                         */
-    PRINT_ALL           /**< Print everything possible */
-  };
-
   /**
-   * @brief print using PrintPreset
-   *
-   * @param out oustream
-   * @param preset preset to be used
+   * @brief     set Property of layer
+   * @param[in] values values of property
+   * @retval #ML_ERROR_NONE Successful.
+   * @retval #ML_ERROR_INVALID_PARAMETER invalid parameter.
    */
-  virtual void printPreset(std::ostream &out,
-                           PrintPreset preset = PrintPreset::PRINT_SUMMARY) = 0;
+  virtual int setProperty(std::vector<std::string> values) = 0;
 };
 
 /**
@@ -234,6 +130,26 @@ std::unique_ptr<Layer>
 createLayer(const std::string &type,
             const std::vector<std::string> &properties = {});
 
+/**
+ * @brief General Layer Factory function to register Layer
+ *
+ * @param props property representation
+ * @return std::unique_ptr<ml::train::Layer> created object
+ */
+template <typename T,
+          std::enable_if_t<std::is_base_of<Layer, T>::value, T> * = nullptr>
+std::unique_ptr<Layer> createLayer(const std::vector<std::string> &props = {}) {
+  std::unique_ptr<Layer> ptr = std::make_unique<T>();
+
+  if (ptr->setProperty(props) != ML_ERROR_NONE) {
+    throw std::invalid_argument("Set properties failed for layer");
+  }
+  return ptr;
+}
+
+/**
+ * Aliases for various layers and losses
+ */
 namespace layer {
 
 /**
@@ -398,40 +314,6 @@ std::unique_ptr<Layer>
 CrossEntropy(const std::vector<std::string> &properties = {});
 
 } // namespace loss
-
-/**
- * @brief General Layer Factory function to register Layer
- *
- * @param props property representation
- * @return std::unique_ptr<ml::train::Layer> created object
- */
-template <typename T,
-          std::enable_if_t<std::is_base_of<Layer, T>::value, T> * = nullptr>
-std::unique_ptr<Layer> createLayer(const std::vector<std::string> &props = {}) {
-  std::unique_ptr<Layer> ptr = std::make_unique<T>();
-
-  if (ptr->setProperty(props) != ML_ERROR_NONE) {
-    throw std::invalid_argument("Set properties failed for layer");
-  }
-  return ptr;
-}
-
-using CreateLayerFunc = Layer *(*)();
-using DestroyLayerFunc = void (*)(Layer *);
-
-/**
- * @brief  Layer Pluggable struct that enables pluggable layer
- *
- */
-typedef struct {
-  CreateLayerFunc createfunc;   /**< create layer function */
-  DestroyLayerFunc destroyfunc; /**< destory function */
-} LayerPluggable;
-
-/**
- * @brief pluggable layer must have this structure defined
- */
-extern "C" LayerPluggable ml_train_layer_pluggable;
 
 } // namespace train
 } // namespace ml
